omgggg

+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: suozkara <suozkara@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/23 16:22:58 by suozkara          #+#    #+#             */
-/*   Updated: 2024/12/17 17:27:59 by suozkara         ###   ########.fr       */
+/*   Updated: 2024/12/17 17:35:20 by suozkara         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,7 +20,7 @@ int	ft_putchar(char c);
 int	ft_putstr(char *s);
 int	ft_putnbr(int n);
 int	ft_putuns(unsigned int n);
-int	ft_puthex(unsigned int n, char c);
+int	ft_puthex(unsigned long n, char c);
 int	ft_putptr(unsigned long n);
 int	ft_printf(const char *s, ...);
