+++ ft_printf_u.c
@@ -6,7 +6,7 @@
 /*   By: suozkara <suozkara@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/23 16:30:12 by suozkara          #+#    #+#             */
-/*   Updated: 2024/12/17 17:22:15 by suozkara         ###   ########.fr       */
+/*   Updated: 2024/12/17 17:35:07 by suozkara         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -60,7 +60,7 @@ int	ft_putuns(unsigned int n)
 	return (ct);
 }
 
-int	ft_puthex(unsigned int n, char c)
+int	ft_puthex(unsigned long n, char c)
 {
 	int	ct;
